Important fix! Corrected angle for FOC startup routine.
Previously the angle was set to 90 degree advance...which is correct for
driving the motor at max torque, but the control system is configured to
drive 90 degrees ahead of THAT, so the actual angle was 180 degrees
ahead of the rotor! Zero torque produced, lots of wasted amps. This has
now been corrected and the startup is as smooth as could be.