Added current nulling at motor startup, and stuck throttle detection.
Sums up a whole bunch of samples at 50% duty cycle on all three phases.
This is okay as long as the motor is stopped. If it's rolling when 50%
duty is applied, it will regen brake and the current won't be zero.
That's why this new routine is skipped above a (very low) motor speed
threshold.

Stuck throttle detection checks if the throttle position is above the
minimum at startup. If it is, the throttle needs to decrease below the
minimum for some duration before any throttle output will be accepted.