Fix to current sensing, minor fixes to power calculations.
Changed the "choose the lowest two currents" to decide based on PWM duty
cycle rather than measured current. This is more reliable in case the
measured current is noisy. Power calculations were giving too high phase
current before, since it was applying a factor of 1.5 to the current.
This is a correct factor to use for the total power, but not the phase
current. Has been corrected.